Fixed pretest script

+++ test/fixtures/pretest.js
@@ -1,5 +1,6 @@
 var fs = require('fs');
 
-fs.unlinkSync(__dirname + '/Ohm.db');
+if(fs.existsSync(__dirname + '/Ohm.db'))
+  fs.unlinkSync(__dirname + '/Ohm.db');
 fs.createReadStream(__dirname + '/Test.db')
   .pipe(fs.createWriteStream(__dirname + '/Ohm.db'));
